corrections on US13

+++ src/main/java/MATDISC/US013/Edge.java
@@ -6,11 +6,11 @@ public class Edge {
 
     private Point p2;
 
-    private double price;
+    private int price;
 
-    private static final double PRICE_PER_OMISSION=0;
+    private static final int PRICE_PER_OMISSION=0;
 
-    public Edge(Point p1, Point p2, double price){
+    public Edge(Point p1, Point p2, int price){
         this.p1=p1;
         this.p2=p2;
         if(price>0){
@@ -21,7 +21,7 @@ public class Edge {
         }
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         if(price>0){
             this.price=price;
         }else {
@@ -29,7 +29,7 @@ public class Edge {
             System.out.println("Preço por omissao introduzido");
         }
     }
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
     public Point getP1() {
@@ -41,6 +41,6 @@ public class Edge {
 
     @Override
     public String toString(){
-        return String.format("Edge: %s to %s Cost: %.2f", p1, p2, price);
+        return String.format("Edge: %s to %s Cost: %d", p1, p2, price);
     }
 }
